requirements.txt
asgiref==3.6.0
Django==4.1.7
django-autoslug==1.9.8
sqlparse==0.4.3
